+++ contracts/nois-gateway/src/lib.rs
@@ -1,3 +1,4 @@
+mod attributes;
 mod drand_archive;
 mod job_id;
 mod request_router;